Fixed EPS type from long to double

+++ spark/dpl-interpreter/src/main/java/com/teragrep/pth_07/performance/metric/Eps.java
@@ -81,7 +81,7 @@ public class Eps implements Stubable, PerformanceMetric {
 
     @Override
     public DataType type() {
-        return DataTypes.LongType;
+        return DataTypes.DoubleType;
     }
 
     @Override
